Adding a listing for "The Ghost Map: The Story of London's Most Terrifying Epidemic--and How It Changed Science, Cities, and the Modern World" (#8)
This book is essential reading for anyone that cares about cities, urban
infrastructure and engineering, cartography, open data, and civic
hacking. Published in 2007, it landed just as open data was starting to
take hold in cities across the country and new tools for developing maps
based on this data were becoming readily available. And it accurately
predicted the way that these new tools would change cities.

It tells the story of John Snow and his efforts to identify the source
of a vicious Cholera outbreak in mid-19th Century London. Snow can be
considered, in my opinion, the first civic hacker. A doctor by trade, he
conducted his own surveys and data collection tactics to develop
innovative visualizations that provided different ways of looking at the
outbreak. He used this data and his new maps to correctly point to the
source of the Cholera outbreak - the water supply - and forever after
influenced data science and urban infrastructure.

Excited to add this to your list.